tests

+++ code/Server.Api/Controllers/CourseSubscriptionController.cs
@@ -94,12 +94,13 @@ namespace Server.Api.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteCourseSubscription(int id)
         {
-            CourseSubscription sub = await _courseSubscriptionRepository.GetAsync(id);
+            var existing = await _courseSubscriptionRepository.GetAsync(id);
+            if (existing == null) return NotFound();
             ClaimsPrincipal currentUser = HttpContext.User;
             if (!currentUser.HasClaim(c => c.Type == "username")) return Unauthorized();
             string userEmail = currentUser.Claims.FirstOrDefault(c => c.Type == "username").Value;
             User user = await _userManager.FindByEmailAsync(userEmail);
-            await _hubContext.Groups.RemoveFromGroupAsync(UserHandler.ConnectedIds[user.Id], "Course " + sub.subscribedItemId.ToString());
+            await _hubContext.Groups.RemoveFromGroupAsync(UserHandler.ConnectedIds[user.Id], "Course " + existing.subscribedItemId.ToString());
             await _courseSubscriptionRepository.DeleteAsync(id);
             return NoContent();
         }
